+++ bonus/client_bonus.c
@@ -1,28 +1,46 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   client.c                                           :+:      :+:    :+:   */
+/*   client_bonus.c                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: masoares <masoares@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/02 15:36:20 by masoares          #+#    #+#             */
-/*   Updated: 2023/11/10 20:37:29 by masoares         ###   ########.fr       */
+/*   Updated: 2023/11/10 20:22:32 by masoares         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "minitalk.h"
+#include "../minitalk_bonus.h"
 
 void	null_send(int pid);
 void	send_char(char c, int pid);
 
+void	handler(int sig, siginfo_t *info, void *c)
+{
+	(void) info;
+	(void) c;
+	if (sig == SIGUSR1)
+	{
+		ft_printf("Message was successfully sent and received");
+		exit(EXIT_SUCCESS);
+	}
+	else if (sig != SIGUSR1)
+		ft_printf("Wrong Signal received: %d", sig);
+}
+
 int	main(int ac, char **av)
 {
-	unsigned char	c;	
-	int				pid;
-	int				i;
-	int				j;
+	unsigned char		c;	
+	int					pid;
+	int					i;
+	int					j;
+	struct sigaction	sig;
 
 	j = 0;
+	sig.sa_sigaction = handler;
+	sig.sa_flags = SA_SIGINFO;
+	if (sigaction(SIGUSR1, &sig, NULL) == -1)
+		exit(EXIT_FAILURE);
 	if (ac == 3)
 	{
 		i = 0;
@@ -30,11 +48,12 @@ int	main(int ac, char **av)
 		while (av[2][j] != '\0')
 		{
 			c = (av[2][j++]);
-			if (c <= 176)
-				send_char(c, pid);
+			send_char(c, pid);
 		}
 		i = 0;
 		null_send(pid);
+		while (1)
+			pause();
 	}
 }
